refactor: clean up middleware
