+++ apps/api/src/middleware/global-rate-limiter.ts
@@ -2,14 +2,13 @@ import type { ErrorSchema } from "$schema";
 import type { Bindings } from "$types/bindings";
 import { DurableObjectStore } from "@hono-rate-limiter/cloudflare";
 import type { KeyData } from "@packages/key-types";
-import type { Context } from "hono";
 import { rateLimiter } from "hono-rate-limiter";
 import { createMiddleware } from "hono/factory";
 
 const MILLISECONDS_PER_HOUR = 60 * 60 * 1_000;
 const REQUESTS_PER_HOUR = 25_000;
 
-export const globalRateLimiter = createMiddleware((c: Context<{ Bindings: Bindings }>, next) =>
+export const globalRateLimiter = createMiddleware<{ Bindings: Bindings }>((c, next) =>
   rateLimiter<{ Bindings: Bindings }>({
     windowMs: MILLISECONDS_PER_HOUR,
     limit: async (c) => {
