+++ apps/api/src/middleware/index.ts
@@ -3,4 +3,6 @@ export * from "./global-rate-limiter";
 export * from "./header-injector";
 export * from "./ip-based-rate-limiter";
 export * from "./key-verifier";
+export * from "./openapi-meta";
 export * from "./production-cache";
+export * from "./reference-og-tag-injector";
