+++ apps/api/src/middleware/ip-based-rate-limiter.ts
@@ -1,14 +1,13 @@
 import type { ErrorSchema } from "$schema";
 import type { Bindings } from "$types/bindings";
 import { DurableObjectStore } from "@hono-rate-limiter/cloudflare";
-import type { Context } from "hono";
 import { rateLimiter } from "hono-rate-limiter";
 import { createMiddleware } from "hono/factory";
 
 const MILLISECONDS_PER_HOUR = 60 * 60 * 1_000;
 const REQUESTS_PER_HOUR = 1_000;
 
-export const ipBasedRateLimiter = createMiddleware((c: Context<{ Bindings: Bindings }>, next) =>
+export const ipBasedRateLimiter = createMiddleware<{ Bindings: Bindings }>((c, next) =>
   rateLimiter<{ Bindings: Bindings }>({
     windowMs: MILLISECONDS_PER_HOUR,
     limit: REQUESTS_PER_HOUR,
