+++ apps/api/src/middleware/production-cache.ts
@@ -1,8 +1,8 @@
 import type { Bindings } from "$types/bindings";
-import type { Context, Next } from "hono";
 import { cache } from "hono/cache";
+import { createMiddleware } from "hono/factory";
 
-export const productionCache =
-  (cacheOptions: Parameters<typeof cache>[0]) =>
-  async (c: Context<{ Bindings: Bindings }>, next: Next) =>
-    c.env.CF_ENV === "prod" ? cache(cacheOptions)(c, next) : await next();
+export const productionCache = (cacheOptions: Parameters<typeof cache>[0]) =>
+  createMiddleware<{ Bindings: Bindings }>(async (c, next) =>
+    c.env.CF_ENV === "prod" ? cache(cacheOptions)(c, next) : await next(),
+  );
